+++ R/RBPSpecificity-package.R
@@ -1,4 +1,4 @@
-#' RBPSpecificity: Analyze RBP Binding Specificity and Mutational Sensitivity
+﻿#' RBPSpecificity: Analyze RBP Binding Specificity and Variation Sensitivity
 #'
 #' @description
 #' The `RBPSpecificity` package provides a suite of tools for researchers
@@ -7,7 +7,7 @@
 #' ImmunoPrecipitation) and RBNS (RNA Bind-n-Seq).
 #'
 #' It allows users to quantify K-mer enrichment within RBP binding sites,
-#' assess the inherent specificity (IS) and the mutational sensitivity (MS) of
+#' assess the inherent specificity (IS) and the variation sensitivity (VS) of
 #' RBP towards the target motifs. The package aims to streamline bioinformatic
 #' analyses and provide clear visualizations for interpreting RBP specificity.
 #'
@@ -19,8 +19,8 @@
 #'   \item \code{\link{motifEnrichment}}: Calculates K-mer enrichment from peak data...
 #'   \item \code{\link{returnIS}}: Quantifies the inherent specificity (IS) of RBP...
 #'   \item \code{\link{plotIS}}: Visualizes the distribution of motif enrichment scores with annotation...
-#'   \item \code{\link{returnMS}}: Quantifies the mutational sensitivity (MS) of RBP...
-#'   \item \code{\link{plotMS}}: Generates plots to visualize Mutational Sensitivity
+#'   \item \code{\link{returnVS}}: Quantifies the variation sensitivity (VS) of RBP...
+#'   \item \code{\link{plotVS}}: Generates plots to visualize Variation Sensitivity
 #' }
 #'
 #' @author Soon Yi \email{cu.soonyi@gmail.com}
@@ -44,3 +44,4 @@ utils::globalVariables(c(
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
+